refactor(pages): refactoring feed page
Setting the right folder and componentizing elements.
